crypto-trend-judger/crypto-quoter-all/requirements.txt
python-binance
influxdb_client